+++ authlib/admin_oauth/apps.py
@@ -6,4 +6,5 @@ class AdminOAuthConfig(AppConfig):
     label = "admin_oauth"
 
     def ready(self):
+        from authlib import checks as authlib_checks  # noqa: F401,PLC0415
         from authlib.admin_oauth import checks  # noqa: F401,PLC0415