+++ authlib/backends.py
@@ -1,7 +1,7 @@
 from functools import cache
 
 from django.contrib.auth import get_user_model
-from django.contrib.auth.backends import ModelBackend
+from django.contrib.auth.backends import BaseBackend, ModelBackend
 from django.contrib.auth.models import Permission
 from django.core.exceptions import PermissionDenied
 from django.db.models import ObjectDoesNotExist
@@ -27,7 +27,13 @@ def _all_perms():
     return [f"{app_label}.{codename}" for app_label, codename in queryset]
 
 
-class PermissionsBackend(ModelBackend):
+class RolePermissionsBackend(BaseBackend):
+    """Answers permission checks from the user's role, and nothing else.
+
+    Despite living in ``AUTHENTICATION_BACKENDS`` this backend authenticates
+    nobody; it has to be accompanied by one which does.
+    """
+
     def get_user_permissions(self, user, obj=None):
         # ModelBackend can use an optimized variant of this -- we cannot since
         # we don't know what the permission checking callbacks do.